+++ docs/.vitepress/theme/index.js
@@ -1,6 +1,11 @@
 import DefaultTheme from "vitepress/theme";
 import CurrentRelease from "./CurrentRelease.vue";
 import LatestReleaseNotes from "./LatestReleaseNotes.vue";
+import RoadmapBoard from "./RoadmapBoard.vue";
+import FeaturesCompare from "./FeaturesCompare.vue";
+import FeaturesShots from "./FeaturesShots.vue";
+import FaqCards from "./FaqCards.vue";
+import IntegrationFilter from "./IntegrationFilter.vue";
 import "./custom.css";
 
 export default {
@@ -8,5 +13,10 @@ export default {
   enhanceApp({ app }) {
     app.component("CurrentRelease", CurrentRelease);
     app.component("LatestReleaseNotes", LatestReleaseNotes);
-  },
+    app.component("RoadmapBoard", RoadmapBoard);
+    app.component("FeaturesCompare", FeaturesCompare);
+    app.component("FeaturesShots", FeaturesShots);
+    app.component("FaqCards", FaqCards);
+    app.component("IntegrationFilter", IntegrationFilter);
+  }
 };